Make script better

+++ setup.sh
@@ -3,6 +3,9 @@ sudo add-apt-repository ppa:git-core/ppa -y
 sudo apt-get update -y
 sudo apt-get install git -y
 
+# SSH keys
+[ ! -f ~/.ssh/id_rsa ] && ssh-keygen
+
 # git config
 read -p "Enter git email: " git_email
 git config --global user.email $git_email
@@ -25,9 +28,9 @@ sudo swapon /swapfile
 sudo sysctl vm.swappiness=10
 sudo sysctl vm.vfs_cache_pressure=50
 sudo sh <<SCRIPT
-sudo echo -e '\n/swapfile   none    swap    sw    0   0' >> /etc/fstab
-sudo echo -e '\nvm.swappiness=10' >> /etc/sysctl.conf
-sudo echo -e '\nvm.vfs_cache_pressure = 50' >> /etc/sysctl.conf
+sudo echo -e '\n/swapfile   none    swap    sw    0   0\n' >> /etc/fstab
+sudo echo -e 'vm.swappiness=10' >> /etc/sysctl.conf
+sudo echo -e 'vm.vfs_cache_pressure = 50' >> /etc/sysctl.conf
 SCRIPT
 
 # Sets up firewall
@@ -45,17 +48,15 @@ sudo ufw enable
 
 # Install nvm with latest node
 export NVM_DIR=$HOME/.nvm
-export default_nvm_version=5.0
 curl https://raw.githubusercontent.com/creationix/nvm/master/install.sh | sh
 source $HOME/.nvm/nvm.sh
-nvm install $default_nvm_version
-nvm use $default_nvm_version
-echo -e '\nsource ~/.nvm/nvm.sh' >> ~/.bashrc
-echo -e "nvm use $default_nvm_version &> /dev/null" >> ~/.bashrc
+nvm install stable
+nvm use stable
+echo -e '\nsource ~/.nvm/nvm.sh\n' >> ~/.bashrc
+echo -e 'nvm use stable\n' >> ~/.bashrc
 
-# Install trash CLI and replace rm
+# Install trash CLI
 sudo npm install -g trash
-alias rm=trash
 
 # heroku toolbelt
 wget -qO- https://toolbelt.heroku.com/install-ubuntu.sh | sh
@@ -75,25 +76,33 @@ gem install awesome_print
 gem install md2man
 rvm gemset use default
 
-# Installs phantonjs v1.9.8
-sudo apt-get install build-essential chrpath libssl-dev libxft-dev -y
-sudo apt-get install libfreetype6 libfreetype6-dev -y
-sudo apt-get install libfontconfig1 libfontconfig1-dev -y
-cd $HOME
-export PHANTOM_JS="phantomjs-1.9.8-linux-x86_64"
-wget https://bitbucket.org/ariya/phantomjs/downloads/$PHANTOM_JS.tar.bz2
-sudo mv $PHANTOM_JS.tar.bz2 /usr/local/share/
-cd /usr/local/share/
-sudo tar xvjf $PHANTOM_JS.tar.bz2
-sudo ln -sf /usr/local/share/$PHANTOM_JS/bin/phantomjs /usr/local/share/phantomjs
-sudo ln -sf /usr/local/share/$PHANTOM_JS/bin/phantomjs /usr/local/bin/phantomjs
-sudo ln -sf /usr/local/share/$PHANTOM_JS/bin/phantomjs /usr/bin/phantomjs
+# Installs phantomjs v1.9.8
+read -r -p "Install phantomjs? [y/N] " response
+response=${response,,}    # tolower
+if [[ $response =~ ^(yes|y)$ ]]
+    sudo apt-get install build-essential chrpath libssl-dev libxft-dev -y
+    sudo apt-get install libfreetype6 libfreetype6-dev -y
+    sudo apt-get install libfontconfig1 libfontconfig1-dev -y
+    cd $HOME
+    export PHANTOM_JS="phantomjs-1.9.8-linux-x86_64"
+    wget https://bitbucket.org/ariya/phantomjs/downloads/$PHANTOM_JS.tar.bz2
+    sudo mv $PHANTOM_JS.tar.bz2 /usr/local/share/
+    cd /usr/local/share/
+    sudo tar xvjf $PHANTOM_JS.tar.bz2
+    sudo ln -sf /usr/local/share/$PHANTOM_JS/bin/phantomjs /usr/local/share/phantomjs
+    sudo ln -sf /usr/local/share/$PHANTOM_JS/bin/phantomjs /usr/local/bin/phantomjs
+    sudo ln -sf /usr/local/share/$PHANTOM_JS/bin/phantomjs /usr/bin/phantomjs
+    cd $HOME
+fi
 
 # Installs and sets up postgres
-cd $HOME
-sudo apt-get install postgresql postgresql-contrib libpq-dev -y
-sudo -u postgres createuser --superuser root
-sudo -u postgres psql -U postgres -d postgres -c "alter user root with password '';"
+read -r -p "Install phantomjs? [y/N] " response
+response=${response,,}    # tolower
+if [[ $response =~ ^(yes|y)$ ]]
+    sudo apt-get install postgresql postgresql-contrib libpq-dev -y
+    sudo -u postgres createuser --superuser root
+    sudo -u postgres psql -U postgres -d postgres -c "alter user root with password '';"
+fi
 
 # Installs direnv
 cd $HOME
@@ -101,29 +110,25 @@ sudo apt-get install golang -y
 git clone http://github.com/zimbatm/direnv
 cd direnv
 sudo make install
-echo -e '\neval "$(direnv hook bash)"' >> ~/.bashrc
+echo -e '\neval "$(direnv hook bash)"\n' >> ~/.bashrc
 
 # Installs dotfiles
-read -p "Do you want to install Sam's dotfiles? " -n 1 -r
-echo    # (optional) move to a new line
-if [[ $REPLY =~ ^[Yy]$ ]]
-then
-    cd $HOME
-    git clone https://github.com/SamLau95/vm-dotfiles.git
-    ln -sb vm-dotfiles/.profile .
-    ln -sb vm-dotfiles/.bashrc .
-    ln -sb vm-dotfiles/.bash_aliases .
-    ln -sb vm-dotfiles/.pryrc .
-    ln -sb vm-dotfiles/.tmux.conf .
-    ln -sb vm-dotfiles/.inputrc .
-    source ~/.profile
-    source ~/.bashrc
-fi
+cd $HOME
+pip install mackup
 
-# Installs vim config
-read -p "Do you want to install spf13's vim config? " -n 1 -r
-echo    # (optional) move to a new line
-if [[ $REPLY =~ ^[Yy]$ ]]
-then
-    curl http://j.mp/spf13-vim3 -L -o - | sh
-fi
+ZIP_LINK=https://www.dropbox.com/sh/o2jxysfs3rrtslq/AABodkj1wWjgUpmXF8vlBxn_a?dl=1
+wget -O mackup.zip $ZIP_LINK
+mkdir ~/mackup/Mackup
+unzip mackup.zip -d ~/mackup/Mackup
+cat > ~/.mackup.cfg <<EOF
+[storage]
+engine = file_system
+path = mackup
+directory = Mackup
+EOF
+
+mackup restore
+
+# Install zsh
+sudo apt-get install -y zsh
+echo -e '\nhash zsh 2>/dev/null && exec zsh -l \n' >> ~/.profile
